requirements.txt
django
djangorestframework
python-decouple
psycopg2-binary
djangorestframework-simplejwt
pillow
dj-config-url
celery[redis]
django-filter
drf-yasg
